+++ docs/src/data/globalnav.ts
@@ -11,6 +11,30 @@ export const RIGHT_NAV_LINKS = [
     url: 'https://aws.amazon.com/amplify/',
     order: 7,
   },
+];
+
+export const LEFT_NAV_LINKS = [
+  {
+    type: 'DEFAULT',
+    label: 'Docs',
+    url: 'https://docs.amplify.aws/',
+    order: 1,
+  },
+  {
+    type: 'DEFAULT',
+    label: 'Learn',
+    url: 'https://amplify.aws/learn',
+    order: 2,
+  },
+  {
+    type: 'DEFAULT',
+    label: 'UI Library',
+    url: 'https://ui.docs.amplify.aws/',
+    order: 3,
+  },
+];
+
+export const SOCIAL_LINKS = [
   {
     type: 'ICON',
     label: 'Discord',
@@ -24,38 +48,5 @@ export const RIGHT_NAV_LINKS = [
     url: 'https://twitter.com/AWSAmplify',
     order: 9,
     icon: 'TWITTER',
-  },
-];
-
-export const LEFT_NAV_LINKS = [
-  {
-    type: 'DEFAULT',
-    label: 'Sandbox',
-    url: 'https://sandbox.amplifyapp.com/',
-    order: 1,
-  },
-  {
-    type: 'DEFAULT',
-    label: 'Docs',
-    url: 'https://docs.amplify.aws/',
-    order: 2,
-  },
-  {
-    type: 'DEFAULT',
-    label: 'Learn',
-    url: 'https://amplify.aws/learn',
-    order: 3,
-  },
-  {
-    type: 'DEFAULT',
-    label: 'UI Library',
-    url: 'https://ui.docs.amplify.aws/',
-    order: 4,
-  },
-  {
-    type: 'DEFAULT',
-    label: 'Contribute',
-    url: 'https://www.contributetoamplify.com/',
-    order: 5,
   },
 ];