+++ examples/next/pages/ui/components/authenticator/sign-in-federated/index.page.tsx
@@ -1,13 +1,18 @@
 import { Amplify } from 'aws-amplify';
 
-import { withAuthenticator } from '@aws-amplify/ui-react';
+import { Authenticator } from '@aws-amplify/ui-react';
 import '@aws-amplify/ui-react/styles.css';
 
 import awsExports from './aws-exports';
 Amplify.configure(awsExports);
-
-function App({ signOut }) {
-  return <button onClick={signOut}>Sign out</button>;
+export default function AuthenticatorWithFederation() {
+  return (
+    <Authenticator>
+      {({ signOut }) => (
+        <>
+          <button onClick={signOut}>Sign out</button>
+        </>
+      )}
+    </Authenticator>
+  );
 }
-
-export default withAuthenticator(App);