+++ packages/react/jest.config.js
@@ -3,14 +3,14 @@ module.exports = {
   collectCoverageFrom: ['<rootDir>/src/**/*.(ts|tsx)'],
   // ignore coverage for top level "export", PrimitiveCatalog, and style files
   coveragePathIgnorePatterns: [
-    '<rootDir>/src/(index|internal|legacy|PrimitiveCatalog|styles).(ts|tsx)',
+    '<rootDir>/src/(index|internal|PrimitiveCatalog|styles).(ts|tsx)',
   ],
   coverageThreshold: {
     global: {
-      branches: 71,
+      branches: 72,
       functions: 60,
-      lines: 79,
-      statements: 79,
+      lines: 85,
+      statements: 85,
     },
   },
   globals: { 'ts-jest': { tsconfig: 'tsconfig.json' } },